requirements.txt
pbr>=0.5.21,<1.0

six>=1.7.0

Jinja2
libvirt-python
